+++ lib/msal-core/src/UserAgentApplication.ts
@@ -40,7 +40,7 @@ import { Constants,
     libraryVersion,
     TemporaryCacheKeys,
     PersistentCacheKeys,
-    ErrorCacheKeys,
+    ErrorCacheKeys
 } from "./utils/Constants";
 import { CryptoUtils } from "./utils/CryptoUtils";
 
@@ -95,6 +95,8 @@ export interface CacheResult {
  */
 export type ResponseStateInfo = {
     state: string;
+    timestamp: number,
+    method: string;
     stateMatch: boolean;
     requestType: string;
 };
@@ -151,7 +153,8 @@ export class UserAgentApplication {
     // state variables
     private silentAuthenticationState: string;
     private silentLogin: boolean;
-    private redirectCallbacksSet: boolean;
+    private redirectResponse: AuthResponse;
+    private redirectError: AuthError;
 
     // Authority Functionality
     protected authorityInstance: Authority;
@@ -209,14 +212,13 @@ export class UserAgentApplication {
         // Set the Configuration
         this.config = buildConfiguration(configuration);
 
-        // Set the callback boolean
-        this.redirectCallbacksSet = false;
-
         this.logger = this.config.system.logger;
         this.clientId = this.config.auth.clientId;
         this.inCookie = this.config.cache.storeAuthStateInCookie;
 
         this.telemetryManager = this.getTelemetryManagerFromConfig(this.config.system.telemetry, this.clientId);
+
+        AuthorityFactory.setKnownAuthorities(this.config.auth.validateAuthority, this.config.auth.knownAuthorities);
 
         // if no authority is passed, set the default: "https://login.microsoftonline.com/common"
         this.authority = this.config.auth.authority || DEFAULT_AUTHORITY;
@@ -238,8 +240,11 @@ export class UserAgentApplication {
         WindowUtils.checkIfBackButtonIsPressed(this.cacheStorage);
 
         // On the server 302 - Redirect, handle this
-        if (urlContainsHash && !WindowUtils.isInIframe() && !WindowUtils.isInPopup()) {
-            this.handleAuthenticationResponse(urlHash);
+        if (urlContainsHash) {
+            const stateInfo = this.getResponseState(urlHash);
+            if (stateInfo.method === Constants.interactionTypeRedirect) {
+                this.handleRedirectAuthenticationResponse(urlHash);
+            }
         }
     }
 
@@ -256,7 +261,6 @@ export class UserAgentApplication {
     handleRedirectCallback(authCallback: authResponseCallback): void;
     handleRedirectCallback(authOrTokenCallback: authResponseCallback | tokenReceivedCallback, errorReceivedCallback?: errorReceivedCallback): void {
         if (!authOrTokenCallback) {
-            this.redirectCallbacksSet = false;
             throw ClientConfigurationError.createInvalidCallbackObjectError(authOrTokenCallback);
         }
 
@@ -269,12 +273,10 @@ export class UserAgentApplication {
             this.authResponseCallback = authOrTokenCallback as authResponseCallback;
         }
 
-        this.redirectCallbacksSet = true;
-
-        // On the server 302 - Redirect, handle this
-        const cachedHash = this.cacheStorage.getItem(TemporaryCacheKeys.URL_HASH);
-        if (cachedHash) {
-            this.processCallBack(cachedHash, null);
+        if (this.redirectError) {
+            this.authErrorHandler(Constants.interactionTypeRedirect, this.redirectError, this.redirectResponse);
+        } else if (this.redirectResponse) {
+            this.authResponseHandler(Constants.interactionTypeRedirect, this.redirectResponse);
         }
     }
 
@@ -323,8 +325,7 @@ export class UserAgentApplication {
      */
     loginRedirect(userRequest?: AuthenticationParameters): void {
         // validate request
-        const request: AuthenticationParameters = RequestUtils.validateRequest(userRequest, true, this.clientId, Constants.interactionTypeRedirect, this.redirectCallbacksSet);
-
+        const request: AuthenticationParameters = RequestUtils.validateRequest(userRequest, true, this.clientId, Constants.interactionTypeRedirect);
         this.acquireTokenInteractive(Constants.interactionTypeRedirect, true, request,  null, null);
     }
 
@@ -336,8 +337,7 @@ export class UserAgentApplication {
      */
     acquireTokenRedirect(userRequest: AuthenticationParameters): void {
         // validate request
-        const request: AuthenticationParameters = RequestUtils.validateRequest(userRequest, false, this.clientId, Constants.interactionTypeRedirect, this.redirectCallbacksSet);
-
+        const request: AuthenticationParameters = RequestUtils.validateRequest(userRequest, false, this.clientId, Constants.interactionTypeRedirect);
         this.acquireTokenInteractive(Constants.interactionTypeRedirect, false, request, null, null);
     }
 
@@ -531,7 +531,7 @@ export class UserAgentApplication {
 
             this.updateCacheEntries(serverAuthenticationRequest, account, loginStartPage);
 
-            // populate QueryParameters (sid/login_hint/domain_hint) and any other extraQueryParameters set by the developer
+            // populate QueryParameters (sid/login_hint) and any other extraQueryParameters set by the developer
             serverAuthenticationRequest.populateQueryParams(account, request);
 
             // Construct urlNavigate
@@ -609,7 +609,7 @@ export class UserAgentApplication {
      */
     acquireTokenSilent(userRequest: AuthenticationParameters): Promise<AuthResponse> {
         // validate the request
-        const request = RequestUtils.validateRequest(userRequest, false, this.clientId);
+        const request = RequestUtils.validateRequest(userRequest, false, this.clientId, Constants.interactionTypeSilent);
         const apiEvent: ApiEvent = this.telemetryManager.createAndStartApiEvent(request.correlationId, API_EVENT_IDENTIFIER.AcquireTokenSilent, this.logger);
 
         return new Promise<AuthResponse>((resolve, reject) => {
@@ -645,16 +645,16 @@ export class UserAgentApplication {
                 request.correlationId,
             );
 
-            // populate QueryParameters (sid/login_hint/domain_hint) and any other extraQueryParameters set by the developer
+            // populate QueryParameters (sid/login_hint) and any other extraQueryParameters set by the developer
             if (ServerRequestParameters.isSSOParam(request) || account) {
-                serverAuthenticationRequest.populateQueryParams(account, request);
+                serverAuthenticationRequest.populateQueryParams(account, request, null, true);
             }
             // if user didn't pass login_hint/sid and adal's idtoken is present, extract the login_hint from the adalIdToken
             else if (!account && !StringUtils.isEmpty(adalIdToken)) {
                 // if adalIdToken exists, extract the SSO info from the same
                 const adalIdTokenObject = TokenUtils.extractIdToken(adalIdToken);
                 this.logger.verbose("ADAL's idToken exists. Extracting login information from ADAL's idToken ");
-                serverAuthenticationRequest.populateQueryParams(account, null, adalIdTokenObject);
+                serverAuthenticationRequest.populateQueryParams(account, null, adalIdTokenObject, true);
             }
             const userContainedClaims = request.claimsRequest || serverAuthenticationRequest.claimsValue;
 
@@ -1010,9 +1010,6 @@ export class UserAgentApplication {
             authErr = err;
         }
 
-        // remove hash from the cache
-        this.cacheStorage.removeItem(TemporaryCacheKeys.URL_HASH);
-
         try {
             // Clear the cookie in the hash
             this.cacheStorage.clearMsalCookie(stateInfo.state);
@@ -1030,12 +1027,13 @@ export class UserAgentApplication {
                     response.tokenType = ServerHashParamKeys.ID_TOKEN;
                 }
                 if (!parentCallback) {
-                    this.authResponseHandler(Constants.interactionTypeRedirect, response);
+                    this.redirectResponse = response;
                     return;
                 }
             } else if (!parentCallback) {
+                this.redirectResponse = buildResponseStateOnly(accountState);
+                this.redirectError = authErr;
                 this.cacheStorage.resetTempCacheItems(stateInfo.state);
-                this.authErrorHandler(Constants.interactionTypeRedirect, authErr, buildResponseStateOnly(accountState));
                 return;
             }
 
@@ -1048,64 +1046,63 @@ export class UserAgentApplication {
 
     /**
      * @hidden
-     * This method must be called for processing the response received from the STS. It extracts the hash, processes the token or error information and saves it in the cache. It then
-     * calls the registered callbacks in case of redirect or resolves the promises with the result.
+     * This method must be called for processing the response received from the STS if using popups or iframes. It extracts the hash, processes the token or error
+     * information and saves it in the cache. It then resolves the promises with the result.
      * @param {string} [hash=window.location.hash] - Hash fragment of Url.
      */
     private handleAuthenticationResponse(hash: string): void {
         // retrieve the hash
         const locationHash = hash || window.location.hash;
 
-        // Check if the current flow is popup or hidden iframe
-        const iframeWithHash = WindowUtils.getIframeWithHash(locationHash);
-        const popUpWithHash = WindowUtils.getPopUpWithHash(locationHash);
-        const isPopupOrIframe = !!(iframeWithHash || popUpWithHash);
-
         // if (window.parent !== window), by using self, window.parent becomes equal to window in getResponseState method specifically
         const stateInfo = this.getResponseState(locationHash);
 
-        let tokenResponseCallback: (response: AuthResponse, error: AuthError) => void = null;
-
-        this.logger.info("Returned from redirect url");
-        // If parent window is the msal instance which opened the current window (iframe)
-        if (isPopupOrIframe) {
-            tokenResponseCallback = window.callbackMappedToRenewStates[stateInfo.state];
-        } else {
-            // Redirect cases
-            tokenResponseCallback = null;
-            // if set to navigate to loginRequest page post login
-            if (this.config.auth.navigateToLoginRequestUrl) {
-                this.cacheStorage.setItem(TemporaryCacheKeys.URL_HASH, locationHash);
-                if (window.parent === window) {
-                    const loginRequestUrl = this.cacheStorage.getItem(`${TemporaryCacheKeys.LOGIN_REQUEST}${Constants.resourceDelimiter}${stateInfo.state}`, this.inCookie);
-
-                    // Redirect to home page if login request url is null (real null or the string null)
-                    if (!loginRequestUrl || loginRequestUrl === "null") {
-                        this.logger.error("Unable to get valid login request url from cache, redirecting to home page");
-                        window.location.href = "/";
-                    } else {
-                        window.location.href = loginRequestUrl;
-                    }
-                }
-                return;
-            }
-            else {
-                window.location.hash = "";
-            }
-
-            if (!this.redirectCallbacksSet) {
-                // We reached this point too early - cache hash, return and process in handleRedirectCallbacks
-                this.cacheStorage.setItem(TemporaryCacheKeys.URL_HASH, locationHash);
-                return;
-            }
-        }
-
+        const tokenResponseCallback = window.callbackMappedToRenewStates[stateInfo.state];
         this.processCallBack(locationHash, stateInfo, tokenResponseCallback);
 
         // If current window is opener, close all windows
-        if (isPopupOrIframe) {
-            WindowUtils.closePopups();
+        WindowUtils.closePopups();
+    }
+
+    /**
+     * @hidden
+     * This method must be called for processing the response received from the STS when using redirect flows. It extracts the hash, processes the token or error
+     * information and saves it in the cache. The result can then be accessed by user registered callbacks.
+     * @param {string} [hash=window.location.hash] - Hash fragment of Url.
+     */
+    private handleRedirectAuthenticationResponse(hash: string): void {
+        this.logger.info("Returned from redirect url");
+
+        // clear hash from window
+        window.location.hash = "";
+
+        // if (window.parent !== window), by using self, window.parent becomes equal to window in getResponseState method specifically
+        const stateInfo = this.getResponseState(hash);
+
+        // if set to navigate to loginRequest page post login
+        if (this.config.auth.navigateToLoginRequestUrl && window.parent === window) {
+            const loginRequestUrl = this.cacheStorage.getItem(`${TemporaryCacheKeys.LOGIN_REQUEST}${Constants.resourceDelimiter}${stateInfo.state}`, this.inCookie);
+            const currentUrl = UrlUtils.getCurrentUrl();
+
+            // Redirect to home page if login request url is null (real null or the string null)
+            if (!loginRequestUrl || loginRequestUrl === "null") {
+                this.logger.error("Unable to get valid login request url from cache, redirecting to home page");
+                window.location.href = "/";
+                return;
+            } else if (currentUrl !== loginRequestUrl) {
+                // If loginRequestUrl contains a hash (e.g. Angular routing), process the hash now then redirect to prevent both hashes in url
+                if (loginRequestUrl.indexOf("#") > -1) {
+                    this.logger.info("loginRequestUrl contains hash, processing response hash immediately then redirecting");
+                    this.processCallBack(hash, stateInfo, null);
+                    window.location.href = loginRequestUrl;
+                } else {
+                    window.location.href = `${loginRequestUrl}${hash}`;
+                }
+                return;
+            }
         }
+
+        this.processCallBack(hash, stateInfo, null);
     }
 
     /**
@@ -1121,10 +1118,14 @@ export class UserAgentApplication {
         if (!parameters) {
             throw AuthError.createUnexpectedError("Hash was not parsed correctly.");
         }
-        if (parameters.hasOwnProperty("state")) {
+        if (parameters.hasOwnProperty(ServerHashParamKeys.STATE)) {
+            const parsedState = RequestUtils.parseLibraryState(parameters.state);
+
             stateResponse = {
                 requestType: Constants.unknown,
                 state: parameters.state,
+                timestamp: parsedState.ts,
+                method: parsedState.method,
                 stateMatch: false
             };
         } else {
@@ -1136,13 +1137,13 @@ export class UserAgentApplication {
          */
 
         // loginRedirect
-        if (stateResponse.state === this.cacheStorage.getItem(`${TemporaryCacheKeys.STATE_LOGIN}${Constants.resourceDelimiter}${stateResponse.state}`, this.inCookie) || stateResponse.state === this.silentAuthenticationState) { // loginRedirect
+        if (stateResponse.state === this.cacheStorage.getItem(`${TemporaryCacheKeys.STATE_LOGIN}${Constants.resourceDelimiter}${stateResponse.state}`, this.inCookie) || stateResponse.state === this.silentAuthenticationState) {
             stateResponse.requestType = Constants.login;
             stateResponse.stateMatch = true;
             return stateResponse;
         }
         // acquireTokenRedirect
-        else if (stateResponse.state === this.cacheStorage.getItem(`${TemporaryCacheKeys.STATE_ACQ_TOKEN}${Constants.resourceDelimiter}${stateResponse.state}`, this.inCookie)) { // acquireTokenRedirect
+        else if (stateResponse.state === this.cacheStorage.getItem(`${TemporaryCacheKeys.STATE_ACQ_TOKEN}${Constants.resourceDelimiter}${stateResponse.state}`, this.inCookie)) {
             stateResponse.requestType = Constants.renewToken;
             stateResponse.stateMatch = true;
             return stateResponse;
@@ -1408,7 +1409,8 @@ export class UserAgentApplication {
 
             // Generate and cache accessTokenKey and accessTokenValue
             const expiresIn = TimeUtils.parseExpiresIn(parameters[ServerHashParamKeys.EXPIRES_IN]);
-            expiration = TimeUtils.now() + expiresIn;
+            const parsedState = RequestUtils.parseLibraryState(parameters[ServerHashParamKeys.STATE]);
+            expiration = parsedState.ts + expiresIn;
             const accessTokenKey = new AccessTokenKey(authority, this.clientId, scope, clientObj.uid, clientObj.utid);
             const accessTokenValue = new AccessTokenValue(parameters[ServerHashParamKeys.ACCESS_TOKEN], idTokenObj.rawIdToken, expiration.toString(), clientInfo);
 
@@ -1720,7 +1722,7 @@ export class UserAgentApplication {
      */
     getAccountState (state: string) {
         if (state) {
-            const splitIndex = state.indexOf("|");
+            const splitIndex = state.indexOf(Constants.resourceDelimiter);
             if (splitIndex > -1 && splitIndex + 1 < state.length) {
                 return state.substring(splitIndex + 1);
             }
@@ -1876,10 +1878,6 @@ export class UserAgentApplication {
      * @returns {boolean} true/false
      */
     public getLoginInProgress(): boolean {
-        const pendingCallback = this.cacheStorage.getItem(TemporaryCacheKeys.URL_HASH);
-        if (pendingCallback) {
-            return true;
-        }
         return this.cacheStorage.getItem(TemporaryCacheKeys.INTERACTION_STATUS) === Constants.inProgress;
     }