+++ lib/msal-core/src/telemetry/TelemetryUtils.ts
@@ -10,8 +10,10 @@ export const scrubTenantFromUri = (uri: string): String => {
 
     // validate trusted host
     if (!AADTrustedHostList[url.HostNameAndPort.toLocaleLowerCase()]) {
-        // Should this return null or what was passed?
-        return null;
+        // returning what was passed because the library needs to work with uris that are non
+        // AAD trusted but passed by users such as B2C or others.
+        // HTTP Events for instance can take a url to the open id config endpoint
+        return uri;
     }
 
     const pathParams = url.PathSegments;